+++ client/lib/utils.js
@@ -146,20 +146,20 @@ Utils = {
     },
 
     isMemberAdmin: function(yes, no) {
-        var filter = { userId: Meteor.userId(), memberType: 'admin' };
+        var filter = { userId: Meteor.userId(), memberType: 'admin', approved: true };
         return this.isMemberFilter(filter) ? yes : no;
     },
 
     isMemberNormal: function(yes, no) {
-        var filter = { userId: Meteor.userId(), memberType: 'normal' };
+        var filter = { userId: Meteor.userId(), memberType: 'normal', approved: true };
         return this.isMemberFilter(filter) ? yes : no;
     },
 
     isMemberAll: function(yes, no) {
         var filter = { 
             $or: [ 
-                { userId: Meteor.userId(), memberType: 'admin' },
-                { userId: Meteor.userId(), memberType: 'normal' }
+                { userId: Meteor.userId(), memberType: 'admin', approved: true },
+                { userId: Meteor.userId(), memberType: 'normal', approved: true }
             ]
         };
         return this.isMemberFilter(filter) ? yes : no;