+++ client/views/activities/events.js
@@ -1,0 +1,7 @@
+Template.activities.events({
+    'click .action-card': function(event, t) {
+        Utils.Overlay.open({ template: 'cardDetailWindow', data: {
+            cardId: this.cardId
+        }});
+    }
+});