+++ collections/activities.js
@@ -1,5 +1,7 @@
-Activities = new Mongo.Collection('activities');
+var _this = this;
 
+// Collection
+Activities = new Mongo.Collection('activities');
 
 Activities.helpers({
     board: function() {
@@ -7,6 +9,15 @@ Activities.helpers({
     },
     user: function() {
         return Users.findOne(this.userId);
+    },
+    member: function() {
+        return BoardMembers.findOne(this.memberId);
+    },
+    list: function() {
+        return Lists.findOne(this.listId);      
+    },
+    card: function() {
+        return Cards.findOne(this.cardId);      
     }
 });