+++ collections/lists.js
@@ -1,7 +1,5 @@
 Lists  = new Mongo.Collection('lists');
 
-
-// ALLOWS
 Lists.allow({
     insert: function(userId, doc) { 
         return allowIsBoardMember(userId, doc.boardId);
@@ -15,10 +13,9 @@ Lists.allow({
 });
 
 
-// HELPERS
 Lists.helpers({
     cards: function() {
-        return Cards.find({ listId: this._id }, { sort: { sort: 1 }});
+        return Cards.find({ listId: this._id, archived: false }, { sort: { sort: 1 }});
     },
     board: function() {
         return Boards.findOne(this.boardId);
@@ -26,18 +23,36 @@ Lists.helpers({
 });
 
 
-// LISTS BEFORE HOOK INSERT
 Lists.before.insert(function(userId, doc) {
     doc.createdAt = new Date();
     doc.updatedAt = new Date();
     doc.archived = false;
-
-    // userId native set.
     if (!doc.userId) doc.userId = userId;
 });
 
 
-// LISTS BEFORE HOOK UPDATE
 Lists.before.update(function(userId, doc, fieldNames, modifier) {
     modifier.$set.modifiedAt = new Date();
+});
+
+isServer(function() {
+    Lists.after.insert(function(userId, doc) {
+        Activities.insert({
+            activityType: "createList", 
+            boardId: doc.boardId,
+            listId: doc._id,
+            userId: userId
+        });
+    });
+
+    Lists.after.update(function(userId, doc) {
+        if (doc.archived) {
+            Activities.insert({
+                activityType: "archivedList",
+                listId: doc._id,
+                boardId: doc.boardId,
+                userId: userId
+            });
+        }
+    });
 });